Convert generic row to LitElement/TS (#2636)
